Implement refresh button and modularize code
